+++ Core/Core/CourseUpgrade/CourseUpgradeHandler.swift
@@ -30,41 +30,6 @@ public enum UpgradeState {
     case error(UpgradeError)
 }
 
-public protocol CourseUpgradeHandlerProtocol {
-    typealias UpgradeCompletionHandler = (UpgradeState) -> Void
-    
-    func upgradeCourse(
-        sku: String?,
-        mode: UpgradeMode,
-        productInfo: StoreProductInfo?,
-        pacing: String,
-        courseID: String,
-        componentID: String?,
-        screen: CourseUpgradeScreen,
-        completion: UpgradeCompletionHandler?
-    ) async
-    
-    func fetchProduct(sku: String) async throws -> StoreProductInfo
-}
-
-class CourseUpgradeHandlerProtocolMock: CourseUpgradeHandlerProtocol {
-    init() {}
-    func upgradeCourse(
-        sku: String?,
-        mode: UpgradeMode,
-        productInfo: StoreProductInfo?,
-        pacing: String,
-        courseID: String,
-        componentID: String?,
-        screen: CourseUpgradeScreen,
-        completion: UpgradeCompletionHandler?
-    ) async {}
-    
-    func fetchProduct(sku: String) async throws -> StoreProductInfo {
-        StoreProductInfo(price: NSDecimalNumber(value: 1))
-    }
-}
-
 public class CourseUpgradeHandler: CourseUpgradeHandlerProtocol {
     static var ecommereceURL: String = ""
