+++ Core/Core/CourseUpgrade/CourseUpgradeInteractor.swift
@@ -7,19 +7,6 @@
 
 import Foundation
 
-//sourcery: AutoMockable
-public protocol CourseUpgradeInteractorProtocol {
-    func addBasket(sku: String) async throws -> UpgradeBasket
-    func checkoutBasket(basketID: Int) async throws -> CheckoutBasket
-    @discardableResult
-    func fulfillCheckout(
-        basketID: Int,
-        price: NSDecimalNumber,
-        currencyCode: String,
-        receipt: String
-    ) async throws -> FulfillCheckout
-}
-
 public class CourseUpgradeInteractor: CourseUpgradeInteractorProtocol {
     
     private let repository: CourseUpgradeRepositoryProtocol
