+++ Core/Core/CourseUpgrade/Models/StoreProductInfo.swift
@@ -1,0 +1,14 @@
+//
+//  StoreProductInfo.swift
+//  Core
+//
+//  Created by Vadim Kuznetsov on 22.05.24.
+//
+
+import Foundation
+
+public struct StoreProductInfo {
+    var price: NSDecimalNumber
+    var localizedPrice: String?
+    var currencySymbol: String?
+}
